Refactor action boxes

+++ src/js/admin/pages/servers/main.ts
@@ -3,12 +3,30 @@ import { loader } from "../../../general/loader";
 import { handleErrorResponse, infobox, sthWentWrong } from "../../../general/infobox";
 import { buildUrl, removeFormWarnings, showWarnings } from "../../../general/global";
 
-$(document).delegate("#antispam_question_button_add", "click", function() {
-    showActionBox(currentPage, "antispam_question_add");
+function showNotification(message) {
+    var tableContainer = $(".table-structure .table-container");
+    var notificationContainer = $("#notification_container");
+
+    if (!notificationContainer.length) {
+        notificationContainer = $("<div>", {
+            id: "notification_container",
+        });
+        tableContainer.prepend(notificationContainer);
+    }
+
+    var notification = $("<div>", {
+        class: "notification is-success",
+        html: message,
+    });
+    notificationContainer.html(notification as any);
+}
+
+$(document).delegate("#server_button_add", "click", function() {
+    showActionBox(window.currentPage, "server_add");
 });
 
 $(document).delegate(".table-structure .edit_row", "click", function() {
-    showActionBox(currentPage, "antispam_question_edit", {
+    showActionBox(window.currentPage, "server_edit", {
         id: $(this)
             .closest("tr")
             .find("td[headers=id]")
@@ -16,15 +34,57 @@ $(document).delegate(".table-structure .edit_row", "click", function() {
     });
 });
 
-// Delete antispam question
+$(document).delegate(".table-structure .regenerate-token", "click", function() {
+    var rowId = $(this).closest("tr");
+    var serverId = rowId.children("td[headers=id]").text();
+    var serverName = rowId.children("td[headers=name]").text();
+
+    var confirmText =
+        "Na pewno chcesz wygenerować nowy token dla serwera:\n(" +
+        serverId +
+        ") " +
+        serverName +
+        " ?";
+    if (!confirm(confirmText)) {
+        return;
+    }
+
+    loader.show();
+    $.ajax({
+        type: "POST",
+        url: buildUrl("/api/admin/servers/" + serverId + "/token"),
+        complete: function() {
+            loader.hide();
+        },
+        success: function(content) {
+            if (!content.return_id) {
+                return sthWentWrong();
+            }
+
+            if (content.return_id === "ok") {
+                showNotification("Nowy token: " + content.data.token);
+            }
+
+            infobox.showInfo(content.text, content.positive);
+        },
+        error: handleErrorResponse,
+    });
+});
+
 $(document).delegate(".table-structure .delete_row", "click", function() {
     var rowId = $(this).closest("tr");
-    var antispamQuestionId = rowId.children("td[headers=id]").text();
+    var serverId = rowId.children("td[headers=id]").text();
+    var serverName = rowId.children("td[headers=name]").text();
+    var confirmText = "Na pewno chcesz usunąć serwer:\n(" + serverId + ") " + serverName + " ?";
+
+    if (!confirm(confirmText)) {
+        return;
+    }
 
     loader.show();
     $.ajax({
         type: "DELETE",
-        url: buildUrl(`/api/admin/antispam_questions/${antispamQuestionId}`),
+        url: buildUrl("/api/admin/servers/" + serverId),
         complete: function() {
             loader.hide();
         },
@@ -37,6 +97,7 @@ $(document).delegate(".table-structure .delete_row", "click", function() {
                 // Delete row
                 rowId.fadeOut("slow");
                 rowId.css({ background: "#FFF4BA" });
+
                 refreshAdminContent();
             }
 
@@ -46,14 +107,13 @@ $(document).delegate(".table-structure .delete_row", "click", function() {
     });
 });
 
-// Add antispam question
-$(document).delegate("#form_antispam_question_add", "submit", function(e) {
+$(document).delegate("#form_server_add", "submit", function(e) {
     e.preventDefault();
 
     loader.show();
     $.ajax({
         type: "POST",
-        url: buildUrl("/api/admin/antispam_questions"),
+        url: buildUrl("/api/admin/servers"),
         data: $(this).serialize(),
         complete: function() {
             loader.hide();
@@ -66,7 +126,7 @@ $(document).delegate("#form_antispam_question_add", "submit", function(e) {
             }
 
             if (content.return_id === "warnings") {
-                showWarnings($("#form_antispam_question_add"), content.warnings);
+                showWarnings($("#form_server_add"), content.warnings);
             } else if (content.return_id === "ok") {
                 clearAndHideActionBox();
                 refreshAdminContent();
@@ -78,18 +138,18 @@ $(document).delegate("#form_antispam_question_add", "submit", function(e) {
     });
 });
 
-// Edit antispam question
-$(document).delegate("#form_antispam_question_edit", "submit", function(e) {
+// Edit server
+$(document).delegate("#form_server_edit", "submit", function(e) {
     e.preventDefault();
 
-    var antispamQuestionId = $(this)
+    var serverId = $(this)
         .find("[name=id]")
         .val();
 
     loader.show();
     $.ajax({
         type: "PUT",
-        url: buildUrl("/api/admin/antispam_questions/" + antispamQuestionId),
+        url: buildUrl("/api/admin/servers/" + serverId),
         data: $(this).serialize(),
         complete: function() {
             loader.hide();
@@ -102,7 +162,7 @@ $(document).delegate("#form_antispam_question_edit", "submit", function(e) {
             }
 
             if (content.return_id === "warnings") {
-                showWarnings($("#form_antispam_question_edit"), content.warnings);
+                showWarnings($("#form_server_edit"), content.warnings);
             } else if (content.return_id === "ok") {
                 clearAndHideActionBox();
                 refreshAdminContent();
